@@ -14,14 +14,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.2.4",
-    "@angular/common": "17.2.4",
-    "@angular/compiler": "17.2.4",
-    "@angular/core": "17.2.4",
-    "@angular/forms": "17.2.4",
-    "@angular/platform-browser": "17.2.4",
-    "@angular/platform-browser-dynamic": "17.2.4",
-    "@angular/router": "17.2.4",
+    "@angular/animations": "18.2.12",
+    "@angular/common": "18.2.12",
+    "@angular/compiler": "18.2.12",
+    "@angular/core": "18.2.12",
+    "@angular/forms": "18.2.12",
+    "@angular/platform-browser": "18.2.12",
+    "@angular/platform-browser-dynamic": "18.2.12",
+    "@angular/router": "18.2.12",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
     "@popperjs/core": "^2.11.8",
     "ajv": "^6.12.6",
@@ -31,13 +31,13 @@
     "rxjs": "^7.4.0",
     "tether": "^1.4.7",
     "tslib": "^2.3.0",
-    "zone.js": "^0.14.2"
+    "zone.js": "^0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.2.3",
-    "@angular/cli": "^17.2.3",
-    "@angular/compiler-cli": "^17.2.4",
-    "@angular/language-service": "17.2.4",
+    "@angular-devkit/build-angular": "^18.2.12",
+    "@angular/cli": "^18.2.12",
+    "@angular/compiler-cli": "^18.2.12",
+    "@angular/language-service": "18.2.12",
     "@types/jasmine": "^3.7.8",
     "@types/jasminewd2": "^2.0.10",
     "@types/node": "^13.13.52",
@@ -57,6 +57,6 @@
     "replace-in-file": "6.0.0",
     "ts-node": "^8.10.2",
     "tslint": "^6.1.3",
-    "typescript": "5.2.2"
+    "typescript": "5.5.4"
   }
 }
