@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { HomepageComponent } from './homepage.component';
 import { SimpleRecipeListComponent } from '../simple-recipe-list/simple-recipe-list.component';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { FavouritesComponent } from '../favourites/favourites.component'
 import { StatsComponent } from '../stats/stats.component'
 
@@ -14,9 +14,10 @@ describe('HomepageComponent', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ HomepageComponent, SimpleRecipeListComponent, FavouritesComponent, StatsComponent ],
-      imports: [HttpClientModule ], 
-    })
+    declarations: [HomepageComponent, SimpleRecipeListComponent, FavouritesComponent, StatsComponent],
+    imports: [],
+    providers: [provideHttpClient(withInterceptorsFromDi())]
+})
     .compileComponents();
   }));
 
