@@ -1,15 +1,15 @@
 import { TestBed, inject } from '@angular/core/testing';
 
 import { RecipeService } from './recipe.service';
-import { HttpClient, HttpClientModule, HttpHandler } from '@angular/common/http';
+import { HttpClient, HttpHandler, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('RecipeService', () => {
   beforeEach(() => {
      let service: RecipeService;
      TestBed.configureTestingModule({
-      imports: [HttpClientModule],
-      providers: [RecipeService, HttpClient, HttpHandler]
-     });
+    imports: [],
+    providers: [RecipeService, HttpClient, HttpHandler, provideHttpClient(withInterceptorsFromDi())]
+});
      service = TestBed.get(RecipeService);
   });
 
